requirements.txt
absl-py==0.15.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
contextlib2==21.6.0
cycler==0.10.0
decorator @ file:///tmp/build/80754af9/decorator_1632776554403/work
efficientnet-pytorch==0.7.1
fonttools==4.25.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
grpcio==1.41.0
idna==3.3
imageio==2.9.0
imgaug==0.4.0
ipython @ file:///tmp/build/80754af9/ipython_1598883837425/work
jedi @ file:///tmp/build/80754af9/jedi_1611333758854/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1633637554808/work
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282414123/work
Markdown==3.3.4
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1634667019719/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
ml-collections==0.1.0
munkres==1.1.4
networkx==2.6.3
numpy==1.21.2
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.4.58
pandas @ file:///tmp/build/80754af9/pandas_1602088128026/work
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.18.1
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1619628580290/work
scipy @ file:///tmp/build/80754af9/scipy_1616700973326/work
Shapely==1.7.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1633102299089/work
tifffile==2021.10.12
torch==1.5.1
torchvision==0.6.1
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1632160078689/work
traitlets @ file:///tmp/build/80754af9/traitlets_1632522747050/work
urllib3==1.26.7
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==2.0.2
yacs==0.1.6
